+++ resources/views/cp/tax-rates/index.blade.php
@@ -5,20 +5,22 @@
 <div class="flex items-center justify-between mb-3">
     <h1 class="flex-1">Tax Rates</h1>
 
-    <dropdown-list class="inline-block">
-        <template v-slot:trigger>
-            <button class="button btn-primary flex items-center pr-2">
-                {{ __('Create Tax Rate') }}
-                <svg-icon name="chevron-down-xs" class="w-2 ml-1" />
-            </button>
-        </template>
+    @if(auth()->user()->can('create tax rates'))
+        <dropdown-list class="inline-block">
+            <template v-slot:trigger>
+                <button class="button btn-primary flex items-center pr-2">
+                    {{ __('Create Tax Rate') }}
+                    <svg-icon name="chevron-down-xs" class="w-2 ml-1" />
+                </button>
+            </template>
 
-        @foreach ($taxCategories as $taxCategory)
-            <dropdown-item
-                redirect="{{ cp_route('simple-commerce.tax-rates.create', ['taxCategory' => $taxCategory->id()]) }}"
-            >{{ $taxCategory->name() }}</dropdown-item>
-        @endforeach
-    </dropdown-list>
+            @foreach ($taxCategories as $taxCategory)
+                <dropdown-item
+                    redirect="{{ cp_route('simple-commerce.tax-rates.create', ['taxCategory' => $taxCategory->id()]) }}"
+                >{{ $taxCategory->name() }}</dropdown-item>
+            @endforeach
+        </dropdown-list>
+    @endif
 </div>
 
 @if ($taxRates->count())
@@ -49,16 +51,21 @@
                         </td>
                         <td class="flex justify-end">
                             <dropdown-list class="mr-1">
-                                <dropdown-item :text="__('Edit')" redirect="{{ $taxRate->editUrl() }}"></dropdown-item>
-                                <dropdown-item :text="__('Delete')" class="warning" @click="$refs.deleter.confirm()">
-                                    <resource-deleter
-                                        ref="deleter"
-                                        resource-title="{{ $taxRate->name() }}"
-                                        route="{{ $taxRate->deleteUrl() }}"
-                                        :reload="true"
-                                        @deleted="document.getElementById('taxRate_{{ $taxRate->id() }}').remove()"
-                                    ></resource-deleter>
-                                </dropdown-item>
+                                @if(auth()->user()->can('edit tax rates'))
+                                    <dropdown-item :text="__('Edit')" redirect="{{ $taxRate->editUrl() }}"></dropdown-item>
+                                @endif
+
+                                @if(auth()->user()->can('delete tax rates'))
+                                    <dropdown-item :text="__('Delete')" class="warning" @click="$refs.deleter.confirm()">
+                                        <resource-deleter
+                                            ref="deleter"
+                                            resource-title="{{ $taxRate->name() }}"
+                                            route="{{ $taxRate->deleteUrl() }}"
+                                            :reload="true"
+                                            @deleted="document.getElementById('taxRate_{{ $taxRate->id() }}').remove()"
+                                        ></resource-deleter>
+                                    </dropdown-item>
+                                @endif
                             </dropdown-list>
                         </td>
                     </tr>
